% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedzaiBrushableHistogram.R
\name{feedzaiBrushableHistogramInput}
\alias{feedzaiBrushableHistogramInput}
\title{<Add Title>}
\usage{
feedzaiBrushableHistogramInput(inputId, default = "")
}
\description{
<Add Description>
}
