% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedzaiBrushableHistogram.R
\name{updateFeedzaiBrushableHistogramInput}
\alias{updateFeedzaiBrushableHistogramInput}
\title{<Add Title>}
\usage{
updateFeedzaiBrushableHistogramInput(session, inputId, value,
  configuration = NULL)
}
\description{
<Add Description>
}
